requirements.txt
ansi2html==1.8.0
asttokens==2.2.1
backcall==0.2.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.3
dash==2.18.2
dash-bootstrap-components==1.4.1
dash-bootstrap-templates==1.3.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
Flask==2.2.5
idna==3.4
importlib_metadata==8.5.0
ipykernel==6.23.3
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.3.0
jupyter_core==5.3.1
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.25.0
packaging==23.1
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.8.0
plotly==5.15.0
prompt-toolkit==3.0.38
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
Pygments==2.15.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.1
pytz==2023.3
pyzmq==25.1.0
requests==2.31.0
retrying==1.3.4
six==1.16.0
stack-data==0.6.2
tenacity==8.2.2
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.7.0
tzdata==2023.3
urllib3==2.0.3
wcwidth==0.2.6
Werkzeug==2.2.3
zipp==3.21.0
